started on a charred pumpkin pastures biome and some other things

+++ src/main/java/j0sh/javadungeons/fluids/DungeonsWaterFluid.java
@@ -41,7 +41,7 @@ public abstract class DungeonsWaterFluid extends DungeonsBaseFluid {
         
         @Override
         public int getLevel(FluidState state) {
-           return (Integer)state.get(LEVEL);
+           return state.get(LEVEL);
         }
         
         @Override
